dispatcher/requirements.txt
anyio==3.7.1
email-validator==2.0.0.post2
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.18.0
idna==3.4
iniconfig==2.0.0
pydantic_core==2.10.1
sentinels==1.0.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
tomli==2.0.1
fastapi-scheduler
redis
pymongo
pydantic[dotenv]
python-decouple
fastapi==0.103.2
pydantic[email]
uvicorn==0.17.4
limits==3.7.0
aio-pika
